requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.28.84
botocore==1.31.84
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
exceptiongroup==1.1.3
fastapi==0.104.1
flower==1.2.0
h11==0.14.0
humanize==4.8.0
idna==3.4
jmespath==1.0.1
jose==1.0.0
kombu==5.3.3
packaging==23.2
prometheus-client==0.18.0
prompt-toolkit==3.0.40
psycopg2==2.9.9
pydantic==2.4.2
pydantic-core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==4.3.4
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tornado==6.3.3
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.9
wrapt==1.16.0
